prediction-service/requirements.txt
# LLM Prediction Service Dependencies

# Google Gemini
google-generativeai==0.3.2

# FastAPI (for prediction service API)
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation
pydantic==2.5.3

# HTTP Client
httpx==0.26.0

# Environment
python-dotenv==1.0.0

# Utilities
tenacity==8.2.3
tiktoken==0.5.2

# Database Client (to query race data)
asyncpg==0.29.0
sqlalchemy==2.0.25

# Caching
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
